A gpart now only accepts M2P interactions if the cell it is interacting with has 'enough' particles.
For now this number is the multipole order squared (although Dehnen recomends 4
times this number). Tests will have to be done to see at what point this starts
to slow down the computation.

The idea is that if the number of particles in a cell is small enough, doing
M2L uses the same amount (or more) operations that it would take just doing
P-P.

Also, P-P will always be more acurate.