When invoke_stf is enabled, the indexing will always follow the stf_count.
So all stf outputs will increase by 1 and be chronological (0,1,2,3,...n)

The snapshots will also be chronological but will increase by arbitary amounts,
they will share the number of the stf output time they relate to.

So the stfs could go (0,1,2,3,4,5,6,7,8,9) giving us 10 group cataloges. But
there are only 3 snapshots, which could be at any of these number (1,7,8) for
example.

The behaviour when not invoking stf is unchanged. The snapshots and stf outputs
follow their own counts.