If you let two cells with few particles interact directly (i.e, Ni*Nj < e->gravity_properties->min_ij_M2L goes P-P).
And if you have a high multipole order.

And you have a large cell count asymmetry (like Ni=1 and Nj=100).

If min_ij_M2L is more than the split size, the gravity_caches wont be big
enough.

So I've changed the size of the cache to be the maximum between the split size
and min_ij_M2L.